pyproject.toml
[tool.poetry]
name = "ampel-ztf"
version = "0.7.1-alpha.17"
description = "Zwicky Transient Facility support for the Ampel system"
authors = [
    "Valery Brinnel",
    "Jakob van Santen <jakob.van.santen@desy.de>",
    "Sjoert van Velzen",
    "Jakob Nordin",
]
maintainers = [
    "Jakob van Santen <jakob.van.santen@desy.de>"
]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://ampelproject.github.io"
repository = "https://github.com/AmpelProject/Ampel-ZTF"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Typing :: Typed",
]
packages = [
    {include = "ampel"}
]
include = [
    'conf/*/*.json',
    'conf/*/*/*.json',
    'conf/*/*.yaml',
    'conf/*/*/*.yaml',
    'conf/*/*.yml',
    'conf/*/*/*.yml',
]

[tool.poetry.dependencies]
ampel-interface = {version = "@0.7.1-alpha.7"}
ampel-core = {version = "@0.7.1-alpha.5"}
ampel-photometry = {version = "~0.7.1-alpha.0"}
ampel-alerts = {version = "~0.7.1-alpha.3"}
ampel-ztf-archive = {version = "~0.7.0-alpha.0", optional = true}
python = "^3.8"
astropy = "^4.2"
matplotlib = "^3.3.4"
aiohttp = "^3.7.3"
nest-asyncio = "^1.4.3"
backoff = "^1.10.0"
confluent-kafka = "^1.5.0"
fastavro = "^1.3.2"
requests = "^2.25.1"
requests-toolbelt = "^0.9.1"
light-curve-python = "^0.2.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
pytest-mock = "^3.5.1"
mongomock = "^3.22.1"
mypy = "^0.800"
pytest-timeout = "^1.4.2"
pytest-asyncio = "^0.14.0"

[tool.poetry.extras]
archive = ["ampel-ztf-archive"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
relative_files = true
